% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_add_variables.R
\name{add_bcf}
\alias{add_bcf}
\title{Estimate bio concentration factor}
\usage{
add_bcf(p, sub_info)
}
\arguments{
\item{p}{Parameter table created with \code{\link[=oneYear_matrix]{oneYear_matrix()}}}

\item{sub_info}{The table containing additional substance information loaded
with \code{\link[=read_fcr_input]{read_fcr_input()}}}
}
\value{
The Parameter table extended by a column for the bio concentration
factor.
}
\description{
Estimation is based on the FCR prepared Monte Carlo table
}
