% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_add_variables.R
\name{add_deposition}
\alias{add_deposition}
\title{Calculate soil mass related daily atmospheric deposition}
\usage{
add_deposition(p)
}
\arguments{
\item{p}{Parameter table created with \code{\link[=oneYear_matrix]{oneYear_matrix()}}}
}
\value{
The Parameter table extended by a column for the atmospheric
Deposition. Unit is mg/(kg * d).
}
\description{
Calculation is based on the FCR prepared Monte Carlo table
}
