% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_add_variables.R
\name{add_DT50}
\alias{add_DT50}
\title{Estimate biological half-life}
\usage{
add_DT50(p, sub_info)
}
\arguments{
\item{p}{Parameter table created with \code{\link[=oneYear_matrix]{oneYear_matrix()}}}

\item{sub_info}{The table containing additional substance information loaded
with \code{\link[=additional_substanc_info]{additional_substanc_info()}}}
}
\value{
The Parameter table extended by a column for the bio concentration
factor.
}
\description{
Estimation is based on the FCR prepared Monte Carlo table
}
\details{
The estimation is based on the K_d value of a substance according to the
technical guidance document on risk assessment Part II (table 8). If there
is a distribution of K_d values, the median K_d is used for the estimation.
In this function there is no distinction between biodegradbility classes.
All substances are assumend to be inherently biodegradable to encourage the
user of further half-life research.
}
