% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_add_variables.R
\name{add_Henry}
\alias{add_Henry}
\title{Estimate Henry Coefficient}
\usage{
add_Henry(p)
}
\arguments{
\item{p}{Parameter table created with \code{\link[=oneYear_matrix]{oneYear_matrix()}}}
}
\value{
The Parameter table extended by a column for Henry coefficient.
Unit is (Pa*m3)/mol
}
\description{
Estimation is based on the FCR prepared Monte Carlo table
}
