% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_add_variables.R
\name{add_Kd}
\alias{add_Kd}
\title{Estimate Sorption Coefficient}
\usage{
add_Kd(p, sub_info)
}
\arguments{
\item{p}{Parameter table created with \code{\link[=oneYear_matrix]{oneYear_matrix()}}}

\item{sub_info}{The table containing additional substance information loaded
with \code{\link[=additional_substanc_info]{additional_substanc_info()}}}
}
\value{
The Parameter table extended by a column for the sorption coefficient
}
\description{
Estimation is based on the FCR prepared Monte Carlo table
}
