% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_add_variables.R
\name{add_kleach}
\alias{add_kleach}
\title{Calculate infiltration rate}
\usage{
add_kleach(p)
}
\arguments{
\item{p}{Parameter table created with \code{\link[=oneYear_matrix]{oneYear_matrix()}}}
}
\value{
The Parameter table extended by a column for the infiltration rate.
Unit is 1/d.
}
\description{
Calculation is based on the FCR prepared Monte Carlo table
}
