% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_add_variables.R
\name{add_kplant}
\alias{add_kplant}
\title{Calculate plant uptake rate}
\usage{
add_kplant(p)
}
\arguments{
\item{p}{Parameter table created with \code{\link[=oneYear_matrix]{oneYear_matrix()}}}
}
\value{
The Parameter table extended by a column for the plant uptake rate.
Unit is 1/d.
}
\description{
Calculation is based on the FCR prepared Monte Carlo table
}
