% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_add_variables.R
\name{add_kvolat}
\alias{add_kvolat}
\title{Calculate volatisation rate}
\usage{
add_kvolat(p)
}
\arguments{
\item{p}{Parameter table created with \code{\link[=oneYear_matrix]{oneYear_matrix()}}}
}
\value{
The Parameter table extended by a column for the partition
coefficient. Unit is 1/d.
}
\description{
Calculation is based on the FCR prepared Monte Carlo table
}
