% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_add_variables.R
\name{add_PNEC_soil}
\alias{add_PNEC_soil}
\title{Estimate Predicted no-effect concentration for soil organisms}
\usage{
add_PNEC_soil(p)
}
\arguments{
\item{p}{Parameter table created with \code{\link[=oneYear_matrix]{oneYear_matrix()}}}
}
\value{
The Parameter table extended by a column for the Predicted no-effect
concentration for soil organisms. Unit is mg/(kg Dry Matter).
}
\description{
Calculation is based on the FCR prepared Monte Carlo table
}
