% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_add_variables.R
\name{add_SoilWater}
\alias{add_SoilWater}
\title{Estimate Soil-Water partition coefficient}
\usage{
add_SoilWater(p)
}
\arguments{
\item{p}{Parameter table created with \code{\link[=oneYear_matrix]{oneYear_matrix()}}}
}
\value{
The Parameter table extended by a column for the partition
coefficient. Unit is m³/m³.
}
\description{
Estimation is based on the FCR prepared Monte Carlo table
}
