% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_add_variables.R
\name{add_variables}
\alias{add_variables}
\title{Query and complementation of input variables}
\usage{
add_variables(p, info)
}
\arguments{
\item{p}{The input data table created with \code{\link[=read_fcr_input]{read_fcr_input()}}}

\item{info}{Additional input data information, created with
\code{\link[=additional_substanc_info]{additional_substanc_info()}}}
}
\value{
The table p exetended by new variable columns needed for the risk assessment
}
\description{
Query and complementation of input variables
}
