% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_load_data.R
\name{additional_substanc_info}
\alias{additional_substanc_info}
\title{Read Additional pollutant information}
\usage{
additional_substanc_info(path, pollutantName)
}
\arguments{
\item{path}{Path of both Excel files}

\item{pollutantName}{Name of Pollutant as in filename}
}
\value{
A table containing the addiational information.
}
\description{
Reads the sheet "additional_infos" from the pollutant Excel file. The
pollutant excel files must be named "'pollutantName'_sheet.xlsx".
}
