% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substance_properties.R
\name{BCF_regression}
\alias{BCF_regression}
\title{Estimate the bio concentration factor in two different ways}
\usage{
BCF_regression(
  constant,
  beta_ph,
  beta_org,
  beta_conc,
  regType = "direct",
  pH,
  org_c,
  conc
)
}
\arguments{
\item{constant}{Log constant in linear Kd regression}

\item{beta_ph, beta_org, beta_conc}{Log coefficient for pH, organic carbon and soil concentration
in log-linear Kd regression}

\item{regType}{Chacter specifying the way the K_d value is calculated. Either
"direct" which means that the regression directly leads to the sorption
coefficient or"indirect" which means the regression estimates a concentration
in pore water and the sorption coefficient is subsequently be calculated by
the quotient of soil and pore water concentration.}

\item{pH}{pH value}

\item{org_c}{Organic carbon content in percent.}

\item{conc}{Soil concentration in mg/kg}
}
\value{
A vector of bio concentration factors without unit. It is defined as
concentration in plant dry matter divided by concentration in soil.
}
\description{
Estimation via direct or indirect logarithmic (log 10?) linear regression
including pH, organic carbon and soil concentration
}
\details{
The constant value, the regression parameters (beta) and the soil
characteristics can be single numeric values or vectors of the same length.
}
