% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_prepare_data.R
\name{create_mcs_input}
\alias{create_mcs_input}
\title{List of Variables into Monte Carlo simulation input table}
\usage{
create_mcs_input(data_list, nFields)
}
\arguments{
\item{data_list}{A list of input variable tables. Each table must have
columns "value_1", "value_2", "distribution", "site_specific", "pID" and
"shift"}

\item{nFields}{Number of Monte Carlo Cycles}
}
\value{
A matrix with all input variables from the list
}
\description{
List of Variables into Monte Carlo simulation input table
}
