% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sinlge_result.R
\name{CumSumSoil}
\alias{CumSumSoil}
\title{Plot cumulative sum of RQ after application and compare it with initial state}
\usage{
CumSumSoil(v0, v, year_x, xmin = 0.001, xmax = 10)
}
\arguments{
\item{v0}{Vector of initial risk quotient distribution}

\item{v}{Vector of final risk quotient distribution of x years}

\item{year_x}{Numarical value of the last year of the risk assessment}

\item{xmin}{Minimum value of x-Axis}

\item{xmax}{Maximumg value of x-Axis}
}
\value{
Cumulative sums are plotted and the underlying table of quantiles of the
initial and final risk quotients
}
\description{
Plot cumulative sum of RQ after application and compare it with initial state
}
