% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_PEC.R
\name{get_c_i_from_PNEC_porewater}
\alias{get_c_i_from_PNEC_porewater}
\title{Calculate the initial concentration in soil for RQ porewater = 1}
\usage{
get_c_i_from_PNEC_porewater(dat, info, nFields)
}
\arguments{
\item{dat}{List with all the input variables. This list is produced by
function \code{\link[=read_fcr_input]{read_fcr_input()}} from the Excel sheets.}

\item{info}{he table containing additional substance information loaded
with \code{\link[=read_fcr_input]{read_fcr_input()}}}

\item{nFields}{Number of Monte Carlo cycles}
}
\value{
Numeric vector of lenght nFields, containig the initial
concentrations in µg/L
}
\description{
The initial soil concentration is calculated so that the average concetrantion
of the first 30 days leads to a PEC_water that equals the PNEC water
}
\details{
The equation is a combination of the definition for PEC soil and
PEC porewater the from TGD, back-calculating to c0.
}
