% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_PEC.R
\name{get_PEC_porewater}
\alias{get_PEC_porewater}
\title{Predicted pore water concentration}
\usage{
get_PEC_porewater(p, d)
}
\arguments{
\item{p}{Parameter table created with \code{\link[=oneYear_matrix]{oneYear_matrix()}} and extended with
\code{\link[=add_variables]{add_variables()}}}

\item{d}{The number of days after fertilizer application. Somewhere between
1 and the number of days of the growing period, since plant uptake is
considered}
}
\value{
Numeric vector of concentrations per field in µg/L
}
\description{
Average predicted environmental concentration (PEC) over d days after
fertilizer application. The Calculation is based on the FCR prepared Monte
Carlo table
}
