% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_interpretation.R
\name{get_risk}
\alias{get_risk}
\title{Risk quotients and high risk scenarios with and without fertilization}
\usage{
get_risk(fertPEC, noFertPEC, PNEC)
}
\arguments{
\item{fertPEC}{A vector of predicted environmental concentrations at the
end of simulation with fertilization}

\item{noFertPEC}{A vector of predicted environmental concentrations at the
end of simulation without fertilization. Both vectors must be based on the
same environmental conditions during simulation.}

\item{PNEC}{A numeric value, defining the predicted no-effect concentration}
}
\value{
Data frame with PECs, Risk Quotient and identification of high-risk scenarios
for both simulations, with and without fertilization, as well as the
risk quotient difference due to fertilization
}
\description{
Risk quotients and high risk scenarios with and without fertilization
}
