% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substance_properties.R
\name{Kd_regression}
\alias{Kd_regression}
\title{Estimate the soil sorption coefficient in two different ways}
\usage{
Kd_regression(
  constant = 0,
  beta_ph = 0,
  beta_org = 0,
  beta_conc = 0,
  regType = "direct",
  pH = 7,
  org_c = 20,
  conc = 1
)
}
\arguments{
\item{constant}{Log constant in linear Kd regression}

\item{beta_ph, beta_org, beta_conc}{Log coefficient for pH, organic carbon and soil concentration
in log-linear Kd regression}

\item{regType}{Chacter specifying the way the K_d value is calculated. Either
"direct" which means that the regression directly leads to the sorption
coefficient or"indirect" which means the regression estimates a concentration
in pore water and the sorption coefficient is subsequently be calculated by
the quotient of soil and pore water concentration.}

\item{pH}{pH value}

\item{org_c}{Organic carbon content in percent.}

\item{conc}{Soil concentration in mg/kg}
}
\value{
A vector of sorption coefficients in L/kg
}
\description{
Estimation via direct or indirect logarithmic linear regression including
pH, organic carbon and soil concentration
}
\details{
The constant value, the regression parameter (beta) and the soil
characteristics can be single numeric values or vectors of the same length.
}
