% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_PEC.R
\name{longterm_PEC}
\alias{longterm_PEC}
\title{PECs for soil, porewater and human consumption over a long period of time}
\usage{
longterm_PEC(
  dat,
  info,
  years,
  nFields,
  use_mixing_factor = FALSE,
  food_only = TRUE,
  growing_period = 180,
  t_res = 1,
  traceBackVariables = FALSE
)
}
\arguments{
\item{dat}{List with all the input variables. This list is produced by
function \code{\link[=read_fcr_input]{read_fcr_input()}} from the Excel sheets.}

\item{info}{he table containing additional substance information loaded
with \code{\link[=additional_substanc_info]{additional_substanc_info()}}}

\item{years}{Years of fertilizer application}

\item{nFields}{Number of Monte Carlo cycles}

\item{use_mixing_factor}{Not working yet! If TRUE, porewater is diluted by pollutant free
groundwater for a more realistic estimation of risks in groundwater.
However the TGD approach assumes porewater = groundwater for the assessment.}

\item{food_only}{If TRUE, the predicted human consumption via food is
multiplied by 2 to compensate the disregard of water consumption}

\item{growing_period}{Numeric value specifiyng the number of days with plant
growth}

\item{t_res}{Temporal resolution to be returned in the output matrix
(no effect on the calculation). The default is 1 (every year) which can lead
to a very large output matrix.}

\item{traceBackVariables}{If TRUE, the variables of every simulation year are
returned. Note: this can lead to very large lists.}
}
\value{
List with
\enumerate{
\item table for all PEC types per year,
\item table for concentration course
in top soil in the predefined temporal resolution,
\item optional: If traceBackVariables TRUE: list of distributed input variables
for each year, if traceBackVariables FALSE: Input variables of the first
year including initial situations for human consumption and
inital concentration in porewater
}
}
\description{
Based on the input data provided in Excel sheets, 3 different PECs are
calculated. The number of years and number of agricultural fields with different
properties (Number of Monte Carlo cycles) can be chosen.
}
