% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_PEC.R
\name{one_year}
\alias{one_year}
\title{Calculation of the pollutant concentation course in top soil}
\usage{
one_year(p, growing_period, t_res)
}
\arguments{
\item{p}{Parameter table created with \code{\link[=oneYear_matrix]{oneYear_matrix()}}}

\item{growing_period}{Numeric value specifiyng the number of days with plant
growth}

\item{t_res}{Temporal resolution to be returned in the output matrix
(no effect on the calculation)}
}
\value{
Matrix with number of columns equal to the rows of p and number of rows
depending on the defined timesteps with t_res
}
\description{
Estimation is based on the FCR prepared Monte Carlo table. The output
rate is split into one rate with plant uptake during plant growth and
one rate without plant uptake during winter time.
}
\details{
A time period can be considered as growing period if 1) plants grow and 2)
plants will be harvested after growing. Otherwise the uptaken pollutants will
return to the soil.
}
