% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_prepare_data.R
\name{oneYear_matrix}
\alias{oneYear_matrix}
\title{Creates a matrix with all Input Data for one year Concentration calculation}
\usage{
oneYear_matrix(dat, c_i, nFields, firstYear = FALSE, use_mixing_factor = FALSE)
}
\arguments{
\item{dat}{List with all the input variables. This list is produced by
function \code{\link[=read_fcr_input]{read_fcr_input()}} from the Excel sheets.}

\item{c_i}{The initial concentration before fertilizer application}

\item{nFields}{The number Monte Carlo Simulations. This can be seen as
different agriculatural fields with different charactersitics.}

\item{firstYear}{We will see}

\item{use_mixing_factor}{If TRUE, a mixing factor for porewater dilution by
unpolluted groundwater is used to get the Risk quotient for Groundwater. The
defahutl is FALSE. In that case pore water concentration is identical with
groundwater concentration as is the approach in the Technical Guidance
Document}
}
\value{
A Matrix with all available input data for the Assessment. The columns
represent the input variabls, the rows correspond to the definied number of
fields aka. number of Monte Carlo cycles.
}
\description{
The information of the Excel input files are used to prepare a Monte Carlo
based data input for the calculation of a yearly concentration dynamic.
}
