% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rderived}
\alias{rderived}
\title{Distribution derived by comparable circumstances}
\usage{
rderived(n, min, max, a = -Inf)
}
\arguments{
\item{n}{Number of samples to be drawn}

\item{min}{Minimum value of normal comparables}

\item{max}{Maximum value of normal comparables}

\item{a}{Minimum of distribution (Default is -Inf)}
}
\value{
numeric vector of length n with randomly drawn values according to
the distribution
}
\description{
This function combines a uniform distribution from minimum to maximum with
the upper and lower tail of a normal distribution.
}
\details{
95\%, 2.5\% and 2.5\% of all drawn values originate from the uniform
distribution, the lower tail, and the upper tail of a normal distribution,
respectively. The normal distribution is characterized by the mean of
minimum and maximum. The standard deviation is derived from the 2.5\% and
97.5\% quantile definied by the minimum and maximum values.
(sd = (max - mean) / 1.959963)
}
