% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdist}
\alias{rdist}
\title{Random samples from specified distribution}
\usage{
rdist(value_1, value_2, n, dist_name, shift = 0, seed = NULL)
}
\arguments{
\item{value_1, value_2}{Distribution parameters (see Details)}

\item{n}{Number of samples to be drawn}

\item{dist_name}{Character vector specifying the Name of the distribution.
Either "none", "uniform", "normal" "tnormal" (for truncated), "lognormal",
"gamma", "derive", "tderive" or "logderive"}

\item{shift}{An numeric value defining a subsequent shift of the distribution
The default is 0 (-> no shift)}

\item{seed}{A numeric value to set the seed for random selection. The default
is NULL -> no seed}
}
\value{
Numeric vector of length n with randomly drawn samples of the
specified distribution
}
\description{
Draw n samples from uniform, normal, truncated normal (-> positive only),
lognormal or gamma distributions based on two input values
}
\details{
The values specified in value_1 and value_2 represent min and max for
"uniform", mean and standard deviation for "normal" and "tnormal", log
mean and log standard deviation for "lognormal" and shape and rate for
"gamma". The shift value is especially interesting for lognormal
or gamma distributions.
}
