% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_load_data.R
\name{read_fcr_input}
\alias{read_fcr_input}
\title{Read environmental and pollutant variables}
\usage{
read_fcr_input(input_path, pollutantName, siteName, fertilizerName = "none")
}
\arguments{
\item{input_path}{Path of the folder "pollutants" and "sites" that contain
the input data excel sheets}

\item{pollutantName}{Name of Pollutant as in filename of the pollutant
(prefix before "_sheet.xslx)}

\item{siteName}{Name of the site as in filename of the environment sheet
(prefix before "_sheet.xslx)}

\item{fertilizerName}{Name of the fertilizer as in filename of the fertilizer
sheet (prefix before "_sheet.xslx). If fertilizerName = "none", fertilizer
application and pollutant concentration will be set to 0.}
}
\value{
List of two items names "dat" and "info" and containing all the
input variable data (information about distribution type, distribution
parameters, type of variability and a unique variable ID) and additional
information of how to simulate substance behavior.
}
\description{
Reads Excel files with environmnetal proporties and substance coefficents.
}
\details{
Both input files must have the same path. The pollutant excel files must be
named "'pollutantName'_sheet.xlsx". Furthermore, both excel tables must have
the same number of columns and column names.
}
