% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_interpretation.R
\name{risk_aggregation}
\alias{risk_aggregation}
\title{Risk Aggregation
Calculation of Delta Risk Quotient and Maximum Risk}
\usage{
risk_aggregation(df_risk)
}
\arguments{
\item{df_risk}{A dataframe created by \code{\link[=get_risk]{get_risk()}}}
}
\value{
A list of 2, the delta risk quotient and the maximum risk
}
\description{
Risk Aggregation
Calculation of Delta Risk Quotient and Maximum Risk
}
\details{
The Delta Risk Quotient (RQ) is the average increase of high-risk scenarios
caused by fertilization. A high-risk scenarios are defined as scenraios with
a final RQ higher than 1. The increase refers to a comparison between two
similar scenarios, the only difference is the application of fertilizer.

The maximum risk is the average RQ of the upper 5\% of the RQ-Distribution.
It is meant to describe the upper tail of the distribution.

Further information can be found in the report of the Eurpean Horizon 2020
project Nextgen "Assessment and risk analysis of NextGen demo case solutions"
(p.151)

\href{https://mp.watereurope.eu/media/publications/D2.1_NextGen_LCA_Risk.pdf}{Nextgen: Deliverable 2.1}
}
