% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_interpretation.R
\name{risk_interpretation}
\alias{risk_interpretation}
\title{Risk Interpretation
Assigns a risk to the aggregated risk parameters}
\usage{
risk_interpretation(delta_RQ, risk_max)
}
\arguments{
\item{delta_RQ}{Numeric vector of length 1: Difference of of high-risk
simulations between scenarios with and without fertilization}

\item{risk_max}{Numeric vector of length 1: Mean value of the Upper five
percent of RQ distribution}
}
\value{
A list of 2, the delta risk quotient and the maximum risk
}
\description{
Risk Interpretation
Assigns a risk to the aggregated risk parameters
}
\details{
Further information can be found in the report of the Eurpean Horizon 2020
project Nextgen "Assessment and risk analysis of NextGen demo case solutions"
(p.151)

\href{https://mp.watereurope.eu/media/publications/D2.1_NextGen_LCA_Risk.pdf}{Nextgen: Deliverable 2.1}
}
