% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rtnorm}
\alias{rtnorm}
\title{Truncate normal distribution}
\usage{
rtnorm(n, mean, sd, a = -Inf, b = Inf)
}
\arguments{
\item{n}{Number of samples to be drawn}

\item{mean}{Mean value of normal distribution}

\item{sd}{Standard deviation of Normal distribution}

\item{a}{Minimum of distribution (Default is -Inf)}

\item{b}{Maximum of distribution (Default is Inf)}
}
\value{
numeric vector of length n with randomly drawn numbers of the
truncated normal distribution
}
\description{
Truncate normal distribution at a minimum or maximum value
}
