% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sinlge_result.R
\name{shadingPlot}
\alias{shadingPlot}
\title{Plot cumulative sum of RQ after application and compare it with initial state}
\usage{
shadingPlot(
  mat_xRow,
  ymin = 0,
  ymax = max(mat_xRow),
  resolution = 0.01,
  xlab = "",
  ylab = "",
  main = ""
)
}
\arguments{
\item{mat_xRow}{A matrix where every row is a distribution for one x value}

\item{ymin}{Minimum value of y-Axis}

\item{ymax}{Maximumg value of y-Axis}

\item{resolution}{The widths of quantiles used for shading (default is 1\%)}

\item{xlab, ylab, main}{Óptional definitions of axis and title}
}
\value{
A plot with overlapping transperant quantiles
}
\description{
Plot cumulative sum of RQ after application and compare it with initial state
}
