% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_concentration_profile.R
\name{temp_c_profile}
\alias{temp_c_profile}
\title{Calculation of the pollutant concentration in top soil}
\usage{
temp_c_profile(conti_input, output_rate, c_i, t_max, t_res, t_beg = 0)
}
\arguments{
\item{conti_input}{Numeric vector of atmospheric depositions in mg/kg topsoil}

\item{output_rate}{Numeric vector of absolute pollutant decay rates.}

\item{c_i}{Numeric vector of initial concentrations in top soil}

\item{t_max}{Numeric value with the last considered timestep (unit
depends on the unit of the decay rate)}

\item{t_res}{Temporal resolution (numeric value) of the timesteps returned
(unit depends on the unit of the decay rate)}

\item{t_beg}{Numeric value of the first considered timestep. This value does
not influence the calculation itself, it is only used for row names.}
}
\value{
Matrix with columns corresponding to the length of the provided
numeric vectors and rows depending on the defined timesteps with
t_max and t_res
}
\description{
Calculation of the pollutant concentration in top soil
}
