% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_add_variables.R
\name{add_Kd}
\alias{add_Kd}
\title{Estimate Sorption Coefficient}
\usage{
add_Kd(p, sub_info)
}
\arguments{
\item{p}{Parameter table created with \code{\link[=oneYear_matrix]{oneYear_matrix()}}. Note: D_air variable
must be available in p if K_d is estimated with linear regression.}

\item{sub_info}{The table containing additional substance information loaded
with \code{\link[=read_fcr_input]{read_fcr_input()}}}
}
\value{
The Parameter table extended by a column for the sorption coefficient
}
\description{
Estimation is based on the FCR prepared Monte Carlo table
}
\details{
Theoretically the initial concentration in soil can become negative if the
pollutant concentration in fertilizes is negative. A negative concentration
makes obviously no sense in a real situation. However, long-term impact is
assessed, a broad concentration range of the pollutant also leads to more
uncertain results. Negative concentrations in one year are averaged out.
For the regression of the sorption coefficient, the concentration must be
positive. Thus, negative values are increased to the concentration in
top soil that would appear after one day of deposition.
}
