% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_load_data.R
\name{read_fcr_input}
\alias{read_fcr_input}
\title{Read environmental and pollutant variables}
\usage{
read_fcr_input(path, pollutantName)
}
\arguments{
\item{path}{Path of both Excel files}

\item{pollutantName}{Name of Pollutant as in filename}
}
\value{
List of 2 tables for environment and pollutant characterstics
}
\description{
Reads Excel files with environmnetal proporties and substance coefficents.
}
\details{
Both input files must have the same path. The pollutant excel files name must
be "'pollutantName'_sheet.xlsc".
}
