% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_load_data.R
\name{read_fcr_input}
\alias{read_fcr_input}
\title{Read environmental and pollutant variables}
\usage{
read_fcr_input(path, pollutantName)
}
\arguments{
\item{path}{Path of both Excel files}

\item{pollutantName}{Name of Pollutant as in filename}
}
\value{
List of one row data frames.
The length equals the number of non NA input variables. Data frames include
information about distribution type, distribution parameters, site specific
character and a unique variable ID
}
\description{
Reads Excel files with environmnetal proporties and substance coefficents.
}
\details{
Both input files must have the same path. The pollutant excel files must be
named "'pollutantName'_sheet.xlsx". Furthermore, both excel tables must have
the same number of columns and column names.
}
